removed unneeded portaudio.lib files and cleaned up to remove some possible loss of data warnings

+++ include/EffectParameters.h
@@ -25,7 +25,7 @@ struct EffectParameters {
 
     //parameters users dont need to see
     std::atomic<float> currentVolume;
-    std::atomic<int> sampleRate;
+    std::atomic<double> sampleRate;
     std::atomic<size_t> delaySamples;
